requirements.txt
django>=6.0.4
djangorestframework>=3.17.1
djangorestframework-simplejwt>=5.5.1
django-filter>=25.2
django-cors-headers>=4.9.0
pillow>=12.2.0
psycopg2-binary>=2.9.11
python-decouple>=3.8
dj-database-url>=3.1.2
